refactor: move offers,products and auth to separated folders
