+++ docs/routes/auth/authentication.yaml
