+++ docs/routes/auth/authorization.yaml
