+++ docs/routes/offers/getOffers.yaml
