+++ docs/routes/products/getProducts.yaml
